analyse/AI Infra/vonLinck Capital/trans/requirements.txt
# Temporal Architecture - Standalone Requirements
# ================================================

# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
joblib>=1.2.0  # For parallel processing (Windows-safe)

# Machine Learning
scikit-learn>=1.0.0
torch>=2.0.0

# Deep Learning Utilities
tqdm>=4.62.0
rich>=13.0.0  # Beautiful terminal UI for dashboard

# Technical Indicators (optional but recommended)
TA-Lib>=0.4.24  # Install with: conda install -c conda-forge ta-lib

# Data Storage
pyarrow>=6.0.0  # For parquet format
h5py>=3.0.0  # For streaming sequence generation (optional but recommended)

# Cloud Storage (optional - only if using GCS)
google-cloud-storage>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.0.0  # MySQL (optional)

# Configuration
python-dotenv>=0.19.0
pyyaml>=6.0

# System Monitoring
psutil>=5.9.0

# API Testing
requests>=2.28.0

# Visualization (optional)
matplotlib>=3.4.0
seaborn>=0.11.0

# Development Tools (optional)
pytest>=7.0.0
jupyter>=1.0.0

# Notes:
# ------
# 1. TA-Lib installation:
#    - Linux/Mac: pip install TA-Lib
#    - Windows: conda install -c conda-forge ta-lib
#    - Or use pure Python fallbacks (slightly slower)
#
# 2. PyTorch installation:
#    - CPU: pip install torch
#    - GPU (CUDA 11.8): pip install torch --index-url https://download.pytorch.org/whl/cu118
#    - GPU (CUDA 12.1): pip install torch --index-url https://download.pytorch.org/whl/cu121
#
# 3. Google Cloud Storage:
#    - Only needed if loading data from GCS
#    - Works without GCS if data is in local cache
